Split up JoinKeyword
