+++ tests/Components/JoinKeywordTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\JoinKeyword;
+use PhpMyAdmin\SqlParser\Components\Lists\JoinKeywords;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
@@ -12,7 +12,7 @@ class JoinKeywordTest extends TestCase
 {
     public function testParseIncomplete(): void
     {
-        $component = JoinKeyword::parse(new Parser(), $this->getTokensList('JOIN a'));
+        $component = JoinKeywords::parse(new Parser(), $this->getTokensList('JOIN a'));
         $this->assertCount(1, $component);
         $this->assertEquals('a', $component[0]->expr->expr);
         $this->assertNull($component[0]->on);
@@ -21,7 +21,7 @@ class JoinKeywordTest extends TestCase
 
     public function testParseIncompleteUsing(): void
     {
-        $component = JoinKeyword::parse(new Parser(), $this->getTokensList('JOIN table2 USING (id)'));
+        $component = JoinKeywords::parse(new Parser(), $this->getTokensList('JOIN table2 USING (id)'));
         $this->assertCount(1, $component);
         $this->assertEquals('table2', $component[0]->expr->expr);
         $this->assertNull($component[0]->on);
@@ -30,7 +30,7 @@ class JoinKeywordTest extends TestCase
 
     public function testBuildAll(): void
     {
-        $component = JoinKeyword::parse(
+        $component = JoinKeywords::parse(
             new Parser(),
             $this->getTokensList(
                 'LEFT JOIN (t2 CROSS JOIN t3 CROSS JOIN t4) ' .
@@ -40,7 +40,7 @@ class JoinKeywordTest extends TestCase
         $this->assertEquals(
             'LEFT JOIN (t2 CROSS JOIN t3 CROSS JOIN t4) ' .
             'ON (t2.a=t1.a AND t3.b=t1.b AND t4.c=t1.c)',
-            JoinKeyword::buildAll($component),
+            JoinKeywords::buildAll($component),
         );
     }
 }
